Cube

Description

This module provides a data type for Rubik's Cube configurations and basic turns of the cube. It provides utilities for working with the cube, as well as an algorithm for solving legal configurations.

Configurations of the Rubik's Cube can be thought of as a permutation of the cubies, as well as an assignment of an orientation to each cubie. The cubies themselves separate into three disjoint sets: the centers, edges, and vertices. There are six center cubies, twelve edge cubies, and eight corner cubies. The center cubies can be oriented in four different ways each, the edge cubies can be oriented in two different ways each, and the vertex cubies can be oriented in three different ways each.

Configurations can compose, and the set of configurations admits a group structure via this composition. Each set of cubies corresponds to a normal subgroup with the structure of a generalized symmetric group.

In total, the Rubik's Cube Group has the structure of \((\mathbb{Z}_4 \wr S_6) \times (\mathbb{Z}_2 \wr S_{12}) \times (\mathbb{Z}_3 \wr S_8)\).

Thus it has order 4^6 * 6! * 2^12 * 12! * 3^8 * 8! = 1530664174762362289520640000.

A configuration of the cube is said to be legal if it can be obtained via a sequence of single turns of the six faces.

Documentation

newtype CubeConfiguration

The main data type for representing configurations of the Rubik's cube.

Internally each cube configuration given by a sextuple made up of three permutations and three tuples.

The permutations follow this enumeration of the cubies in their original position:

(The cube here is displayed as an unwrapped net)

               -----------
              | 2 | 3 | 3 |
              |---+---+---|
              | 2 | 1 | 4 |
              |---+---+---|
              | 1 | 1 | 4 |
               -----------
 -----------   -----------   -----------   -----------
| 2 | 2 | 1 | | 1 | 1 | 4 | | 4 | 4 | 3 | | 3 | 3 | 2 |
|---+---+---| |---+---+---| |---+---+---| |---+---+---|
| 6 | 3 | 5 | | 5 | 2 | 8 | | 8 | 5 | 7 | | 7 | 4 | 6 |
|---+---+---| |---+---+---| |---+---+---| |---+---+---|
| 8 |12 | 5 | | 5 | 9 | 6 | | 6 |10 | 7 | | 7 |11 | 8 |
 -----------   -----------   -----------   -----------
               -----------
              | 5 | 9 | 6 |
              |---+---+---|
              |12 | 6 |10 |
              |---+---+---|
              | 8 |11 | 7 |
               -----------

The tuples give the orientations of the center, edge, and vertex cubies, represented as integers mod 4, mod 2, and mod 3, respectively. The tuples have length 6, 12, and 8, respectively. For centers and corners, a clockwise twist causes an increment to its orientation. Edges can simply be flipped or unflipped.

The value in the nth position in each tuple represents the new position of the 0 face of the nth cubie relative to the position that cubie is now in.

Center cubies only have one face, but this face can be in 4 different orientations. These are enumerated such that a clockwise turn of a center cubie increments its orientation.

Default orientations follow this enumeration of cubie faces in their original orientations:

               -----------
              | 0 | 0 | 0 |
              |---+---+---|
              | 0 | * | 0 |
              |---+---+---|
              | 0 | 0 | 0 |
               -----------
 -----------   -----------   -----------   -----------
| 1 | 1 | 2 | | 1 | 1 | 2 | | 1 | 1 | 2 | | 1 | 1 | 2 |
|---+---+---| |---+---+---| |---+---+---| |---+---+---|
| 1 | * | 1 | | 0 | * | 0 | | 1 | * | 1 | | 0 | * | 0 |
|---+---+---| |---+---+---| |---+---+---| |---+---+---|
| 2 | 1 | 1 | | 2 | 1 | 1 | | 2 | 1 | 1 | | 2 | 1 | 1 |
 -----------   -----------   -----------   -----------
               -----------
              | 0 | 0 | 0 |
              |---+---+---|
              | 0 | * | 0 |
              |---+---+---|
              | 0 | 0 | 0 |
               -----------

Constructors

Cube (CenterP, EdgeP, VertexP, CenterO, EdgeO, VertexO)

Basic turns

data Turn

A data type representing basic turns of the cube. These are the generators of the legal cube group.

Applying each basic turn to the cube in its default state results in a unique cube configuration.

Constructors

I
The Identity (no turn)

U
A single turn of the Up (top) face clockwise

U'
A single turn of the Up (top) face counter-clockwise

F
A single turn of the Front face clockwise

F'
A single turn of the Front face counter-clockwise

L
A single turn of the Left face clockwise

L'
A single turn of the Left face counter-clockwise

B
A single turn of the Back face clockwise

B'
A single turn of the Back face counter-clockwise

R
A single turn of the Right face clockwise

R'
A single turn of the Right face counter-clockwise

D
A single turn of the Down (bottom) face clockwise

D'
A single turn of the Down (bottom) face counter-clockwise

invertTurn :: Turn -> Turn

invertTurns :: [Turn] -> [Turn]

turnToConfig :: Turn -> CubeConfiguration

turnsToConfig :: [Turn] -> CubeConfiguration

Composes a sequence of turns

Displaying configurations

showCubeConfig :: CubeConfiguration -> String

Converts a cube configuration into 6 values, representing the permutations of the center, edge, and vertex cubies, and the orientation tuples of the center, edge, and vertex cubies.

>>> putStrLn $ showCubeConfig $ turnToConfig F
1
(1 8 9 5)
(1 4 6 5)
(0,1,0,0,0,0)
(1,0,0,0,1,0,0,1,1,0,0,0)
(1,0,0,2,2,1,0,0)

This is the default Show implementation for cube configurations.

showCube :: CubeConfiguration -> String

Uses the default assignment of strings to colors

newtype ASCIICube

Extra newtype that allows for an alternate, more visual Show instance for cube configurations

E.g. the identity configuration is displayed as

               -----------
              |   |   |   |
              |---+---+---|
              |   |^ ^|   |
              |---+---+---|
              |   |   |   |
               -----------
 -----------   -----------   -----------   -----------
| X | X | X | |:::|:::|:::| |###|###|###| | o | o | o |
|---+---+---| |---+---+---| |---+---+---| |---+---+---|
| X |^X^| X | |:::|^:^|:::| |###|^#^|###| | o |^o^| o |
|---+---+---| |---+---+---| |---+---+---| |---+---+---|
| X | X | X | |:::|:::|:::| |###|###|###| | o | o | o |
 -----------   -----------   -----------   -----------
               -----------
              | ~ | ~ | ~ |
              |---+---+---|
              | ~ |^~^| ~ |
              |---+---+---|
              | ~ | ~ | ~ |
               -----------

Constructors

ShowCube CubeConfiguration

Basic operations

(|#|) :: TwistyPuzzle a => a -> a -> a infixl 7

Composition of cube configurations

(|#|^) :: (TwistyPuzzle a, Integral b) => a -> b -> a infixl 8

Exponentiation (including negative exponents)

(|#|^|#|) :: TwistyPuzzle a => a -> a -> a infix 8

Conjugation of cube configurations, i.e.,

g |#|^|#| h == h |#|^ (-1) |#| g |#| h

(>|#|<) :: TwistyPuzzle a => a -> a -> a infix 7

Commutator of cube configurations, i.e.,

g >|#|< h == g |#|^ (-1) |#| h |#|^ (-1) |#| g |#| h

Permutation representations

data Cubie

Type for representing individual cubies

Constructors

C Integer Mod4

E Integer Mod2

V Integer Mod3

toPermutation :: CubeConfiguration -> Permutation Cubie

Configurations of the cube can also be seen as permutations of the set of stickers (where the 4 orientations of each center cubie sticker are considered distinct). This manifests as a monomorphism from the group of cube configurations into the permutation group of stickers. Conversely, not every permutation of stickers gives a valid configuration of the cube, for example a vertex sticker can never end up in the place of an edge sticker.

This sends a configuration to a permutation of the stickers, where each sticker is represented as a tuple (X, n, m), where X encodes whether it is a center, edge, or vertex cubie (taking the values C, E, or V, respectively), n represents the cubie the sticker is attached to, and m represents the face of that cubie that the sticker is attached to. For center cubies, m represents the orientation of the sticker. Orientations are not represented as modular integers here for type reasons.

fromPermutation :: Permutation Cubie -> CubeConfiguration

toNumericPermutation :: CubeConfiguration -> Permutation Integer

This sends a configuration to the same permutation of stickers, but with each sticker represented as a number between 1 and 72 (including a different number for each orientation of each center sticker). Center stickers (with their 4 orienations each) take values 1 - 24, edge stickers take values 25 - 48, and vertex stickers take values 49 - 72.

Sticker enumeration:

               -----------
              |50 |27 |51 |
              |---+---+---|
              |26 | 1 |28 |
              |---+---+---|
              |49 |25 |52 |
               -----------
 -----------   -----------   -----------   -----------
|58 |38 |65 | |57 |37 |68 | |60 |40 |67 | |59 |39 |66 |
|---+---+---| |---+---+---| |---+---+---| |---+---+---|
|42 | 3 |41 | |29 | 2 |32 | |44 | 5 |43 | |31 | 4 |30 |
|---+---+---| |---+---+---| |---+---+---| |---+---+---|
|72 |48 |61 | |69 |45 |62 | |70 |46 |63 | |71 |47 |64 |
 -----------   -----------   -----------   -----------
               -----------
              |53 |33 |54 |
              |---+---+---|
              |36 | 6 |34 |
              |---+---+---|
              |56 |35 |55 |
               -----------

fromNumericPermutation :: Permutation Integer -> CubeConfiguration

Group theoretic properties

order :: CubeConfiguration -> Int

Gives the order of an element within the group of configurations (or, equivalently, within the group of sticker permutations)

Solving the cube

isSimilarTo :: CubeConfiguration -> CubeConfiguration -> Bool

Two elements are similar when one can be obtained from the other via a sequence of basic turns This is a congruence relation.

isLegal :: CubeConfiguration -> Bool

An element is legal if it is generated by the basic turns. This is equivalent to being similar to the identity configuration.

generate :: CubeConfiguration -> Maybe [Turn]

Given a legal configuration, returns a sequence of basic turns that produce that configuration

solve :: CubeConfiguration -> Maybe [Turn]

Given a legal configuration, returns a sequence of basic turns that produce the inverse of that configuration

Alternate display

showCubeCustom :: Map Integer String -> CubeConfiguration -> String

Creates the list of string representing the cube configuration as ASCII art Takes as input a Map which sends the numbers 1 - 72 to strings. Each string should be three characteres long. These serve as the stickers for the ASCII art. The stickers are enumerated just as in the definition of toNumericPermutation.