Cube.ClassicCube

Description

The classic Rubik's Cube does not distinguish different orientations for the center cubies

In total, the classic Rubik's Cube Group has the structure of \(S_6 \times (\mathbb{Z}_2 \wr S_{12}) \times (\mathbb{Z}_3 \wr S_8)\).

Thus it has order 6! * 2^12 * 12! * 3^8 * 8! = 373697308291592355840000.

Documentation

makeClassic :: CubeConfiguration -> CubeConfiguration

Ignore center orientations

orderClassic :: CubeConfiguration -> Int

Compute the order of a cube configuration ignoring center orientations

showClassicCubeConfig :: CubeConfiguration -> String

Show a cube configuration ignoring center orientations

data ASCIIClassicCube

What the Cube looks like in its default state:

               -----------
              |   |   |   |
              |---+---+---|
              |   |   |   |
              |---+---+---|
              |   |   |   |
               -----------
 -----------   -----------   -----------   -----------
| X | X | X | |:::|:::|:::| |###|###|###| | o | o | o |
|---+---+---| |---+---+---| |---+---+---| |---+---+---|
| X | X | X | |:::|:::|:::| |###|###|###| | o | o | o |
|---+---+---| |---+---+---| |---+---+---| |---+---+---|
| X | X | X | |:::|:::|:::| |###|###|###| | o | o | o |
 -----------   -----------   -----------   -----------
               -----------
              | ~ | ~ | ~ |
              |---+---+---|
              | ~ | ~ | ~ |
              |---+---+---|
              | ~ | ~ | ~ |
               -----------

showClassicCube :: CubeConfiguration -> String

Create ASCII art for the cube ignoring center orientations