Cube.SudokuCube

Description

This module gives an example of a custom sticker set. The stickers of a sudoku cube are labeled 1-9 such that each face forms a valid sudoku puzzle.

Documentation

data ASCIISudokuCube

What the Cube looks like in its default state:

               -----------
              | 3 | 8 | 7 |
              |---+---+---|
              | 2 | 1 | 9 |
              |---+---+---|
              | 5 | 6 | 4 |
               -----------
 -----------   -----------   -----------   -----------
| 1 | 4 | 2 | | 4 | 3 | 6 | | 9 | 7 | 1 | | 3 | 9 | 7 |
|---+---+---| |---+---+---| |---+---+---| |---+---+---|
| 8 | 7 | 3 | | 7 | 5 | 1 | | 2 | 4 | 8 | | 6 | 1 | 5 |
|---+---+---| |---+---+---| |---+---+---| |---+---+---|
| 6 | 9 | 5 | | 9 | 2 | 8 | | 5 | 3 | 6 | | 8 | 2 | 4 |
 -----------   -----------   -----------   -----------
               -----------
              | 1 | 5 | 2 |
              |---+---+---|
              | 4 | 6 | 3 |
              |---+---+---|
              | 7 | 8 | 9 |
               -----------  

showSudokuCube :: CubeConfiguration -> String