Permutable

Description

A module which implements the generalized symmetric group for a few specific orders.

Documentation

class Foldable z => Permutable z where

Minimal complete definition

(*!), fromList

Methods

(*!) :: Integral b => z a -> b -> a infixl 9

fromList :: [a] -> z a

toPairList :: Integral b => z a -> [(b, a)]

(*?) :: Integral b => z a -> Permutation b -> z a infixr 7

Right action by permutations

(?*) :: Integral b => Permutation b -> z a -> z a infixl 7

Left action by permutations

Tuple Newtypes

newtype Tuple6 a

Constructors

T6 (a, a, a, a, a, a)

newtype Tuple8 a

Constructors

T8 (a, a, a, a, a, a, a, a)

newtype Tuple12 a

Constructors

T12 (a, a, a, a, a, a, a, a, a, a, a, a)

Alternate Constructors

t6 :: a -> a -> a -> a -> a -> a -> Tuple6 a

t8 :: a -> a -> a -> a -> a -> a -> a -> a -> Tuple8 a

t12 :: a -> a -> a -> a -> a -> a -> a -> a -> a -> a -> a -> a -> Tuple12 a