Permutation

Description

A module for working with permutations. A permutation is a bijection from a set onto itself. Permutations can be made from any type, so long as it is an instance of Ord. This is because permutations are stored using Data.Map.

Documentation

data Permutation a

Construction

p :: Ord a => [[a]] -> Permutation a

Construct a permutation from a list of cycles. On invalid input, returns the identity.

pp :: Ord a => [(a, a)] -> Permutation a

Construct a permutation from a list of pairs. On invalid input, returns the identity.

Displaying permutations

showInline :: (Ord a, Show a) => Permutation a -> String

Converts the permutation to a string, showing its cycle decomposition on a single line, i.e.,

(1 2 3)(4 5)

showMultiline :: (Ord a, Show a) => Permutation a -> String

Inserts a newline character between each cycle.

Basic operations and notation for permutations

(?.) :: Ord a => Permutation a -> a -> a infixr 1

Action on an element (Apply the permutation)

>>> p [[1,2,3]] ?. 2
3

(?-) :: Ord a => Permutation a -> [a] -> [a] infixr 1

Action on a set

>>> p [[1,2,3]] ?- [2,4]
[3,4]

(?) :: Ord a => Permutation a -> Permutation a -> Permutation a infixl 7

Multiplication

(o ? q) ?. x == o ?. (q ?. x)

one :: Ord a => Permutation a

The identity (the identity permutation is often written as 1)

one == p [[]]

(?^) :: (Ord a, Integral b) => Permutation a -> b -> Permutation a infixl 8

Exponentiation (including negative exponents)

>>> p [[1,2,3,4,5]] ?^ 2
(1 3 5 2 4)

(?^?) :: Ord a => Permutation a -> Permutation a -> Permutation a infix 8

Conjugation of two permutations, i.e.,

o ?^? q == q ?^ (-1) ? o ? q

(>?<) :: Ord a => Permutation a -> Permutation a -> Permutation a infix 7

Commutator of two perumutations, i.e.,

o >?< q == o ?^ (-1) ? q ?^ (-1) ? o ? q

Useful mathematical functions of permutations

parity :: Ord a => Permutation a -> Mod2

Parity of a permutation

sgn :: (Ord a, Integral b) => Permutation a -> b

Sign of a permutation

order :: Ord a => Permutation a -> Int

Compute the order of a permutation.

subgroupOrder :: Ord a => [Permutation a] -> Int

Compute the order of the group generated by a list of permutations.

cycleOf :: Ord a => a -> Permutation a -> [a]

Find the cycle of the given element in the given permutation.

support :: Ord a => Permutation a -> [a]

The support of a permutation (the set of non-fixed points). Works for permutations created using the supplied constructors.

transpositionDecomposition :: Ord a => a -> Permutation a -> [[a]]

Given z, decompose the permutation into an equivalent product of transpositions of the form (z x). This can be done for any permutation and any z but is not unique.

E.g.,

>>> transpositionDecomposition 1 (p [[1,2,3],[4,5]])
 [[1,3],[1,2],[1,4],[1,5],[1,4]]

threeCycleDecomposition :: Ord a => a -> a -> Permutation a -> [[a]]

Given z1, z2, decompose the permutation into an equivalent product of 3-cycles of the form (z1 z2 x) and (z2 z1 x). This can be done for any even permutation and any z1 and z2 but is not unique. If the permutation is not even, the product will end in a transposition of the form (z1 x).

E.g.,

>>> threeCycleDecomposition 5 6 (p [[1,2],[3,4]])
[[6,5,2],[6,5,1],[5,6,2],[6,5,3],[6,5,1],[6,5,4],[5,6,3]]

Utility functions

fromPairs :: Ord a => [(a, a)] -> Maybe (Permutation a)

toPairs :: Ord a => Permutation a -> [(a, a)]

fromCycles :: Ord a => [[a]] -> Maybe (Permutation a)

toCycles :: Ord a => Permutation a -> [[a]]

toFunction :: Ord a => Permutation a -> a -> a

Functions for testing purposes

permutationOf :: Ord a => [a] -> Gen (Permutation a)

Generate a random permutation of the given list. For use with Test.QuickCheck.